Adding access list support for editors
Editors now need to be in an access list file to actually log in.
This is based on their unique ID that is generated when an editor
is created.

Cleaned up controllers to use res.render if they used res.send.

Removed package-lock.json since it's not a necessary file to track