Initial re-organization (will not run)
Pushing this change in to get the initial bit of reorganization going.
This will not run, since many of the depencies need rewiring/reworking.

Major areas of this change include
- Folding the contents of bin/www into app.js
- Pulling out various steps of initializing the server in app.js into
"loaders"
- Updated the config setup so it uses a config object that can be passed
around that is setup with defaults. This helps prevent needing to check
if an environment variable exists, as it takes advnatage of
short-circuit assiging (e.g. "[default] || [Environment Var]")
- "Services" removes, unifying database CRUD operations to a single
class.
- "Controllers" renamed to "Services" to contain just the business logic
and avoid manipulating Express's req and res objects. This decouples the
business logic from Express