Masssssive clean up and updates
Probably not the best to make a huge commit.

Incorporated a bunch of changes from another project to update how
everything is setup before launching the server. Got rid of the
mongo-dal file since it's redundant and restrictive. Services files were
updated to work without mongo-dal and are now responsible for importing
the schema files. Updated the editor management tool to including
listing editors.